+++ src/include/data/ffm/data.h
@@ -16,17 +16,17 @@ template<typename T>
 class Row {
 
  public:
-  Row(size_t *features, size_t* fields, T* values, int label, T scale, size_t size) : features(features), fields(fields), values(values),
+  Row(int *features, int* fields, T* values, int label, T scale, int size) : features(features), fields(fields), values(values),
                                                                              label(label), scale(scale), size(size) {}
 
   // feature:field:value for all the data points in all the rows
-  size_t* features;
-  size_t* fields;
+  int* features;
+  int* fields;
   T* values;
 
   int label;
   T scale;
-  size_t size;
+  int size;
 
 };
 
@@ -35,7 +35,7 @@ class Dataset {
  public:
   Dataset() {}
 
-  Dataset(int numFields, int numFeatures, size_t numRows, size_t numNodes, size_t *features, size_t* fields, T* values, int *labels, T *scales, size_t *rowPositions) :
+  Dataset(int numFields, int numFeatures, int numRows, int numNodes, int *features, int* fields, T* values, int *labels, T *scales, int *rowPositions) :
       numRows(numRows),
       numNodes(numNodes),
       numFields(numFields),
@@ -48,20 +48,20 @@ class Dataset {
       rowPositions(rowPositions){}
 
   // Number of rows in the dataset
-  size_t numRows;
+  int numRows;
   // Total number of nodes in all the rows
-  size_t numNodes;
+  int numNodes;
   // Total number of fields
   int numFields = 0;
   // Total number of features
   int numFeatures = 0;
 
   // Starting position for each row. Of size nr_rows + 1
-  size_t* rowPositions;
+  int* rowPositions;
 
   // feature:field:value for all the data points in all the rows
-  size_t* features;
-  size_t* fields;
+  int* features;
+  int* fields;
   T* values;
 
   // label and scale for each row
@@ -72,8 +72,8 @@ class Dataset {
   bool cpu = true;
 
   // Total size in bytes required to hold this structure
-  size_t requiredBytes() const {
-    return numNodes * 2 * sizeof(size_t) + numRows * sizeof(T) + numRows * sizeof(int) + numRows * sizeof(T);
+  int requiredBytes() const {
+    return numNodes * 2 * sizeof(int) + numRows * sizeof(T) + numRows * sizeof(int) + numRows * sizeof(T);
   }
 };
